Remove the SPL vector from the PDU vector
We should not encode the 4 byte SPL as the first vector for the PDU.
The PDU should only contain the vectors for the PDU itself and its
payload.
Instead add the SPL vector as part of preparing the vectors that needs to
be sent in smb2_write_to_socket().

We need this separation as at some stage we may want to send multiple PDUs
in a single SPL. I.e. compounding.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>