Add support for sending Compunded commands.
Convert smb2_stat_async to use compound commands to do the open + query + close
in a single roundtrip.
This only works for the sending side since the receive side does not yet
know how to handle compounded responses.

Wireshark does know how to decode both the compunded request as well as
the componded reply that win10 sends back to us. So there is that.

Before this can be checked into master we need to build support to handle
compunded receive side.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>