Add a builtin NTLMSSP module we can use instead of gssapi_krb5
This is a minimal module that only supports NTLMSSP authentication
but not kerberos. However it does not require linking with gssapi
and kerberos thus making it much easier to port where availability
of kerberos+gssapi may be difficult.

To use this module you would configure with :
  ./configure --without-libkrb5
and your appplication must provide proper credentials to the smb2
context before you connect using
  smb2_set_user(smb2, <username>);
  smb2_set_password(smb2, <password>);

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>