Reset the context if connect_share fails
We need to close the fd and reset some fields in the context if the
smb2_connect_share call fails so that we can re-try the connect.

For example if we use the incorrect credentials we might want to
ask the user for the password a second time and re-try to connect
using the same context as before.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>