Signing support for the built-in NTLMSSP mode
For NTLMv2 the session key is to be calculated by the client from
the session base key, which inturn gets generated from the key exchange key.
Since in current implementation we are not doing
key exchange, the key_exchange key is both the session base key
and it is the session key.

Follow - https://blogs.msdn.microsoft.com/openspecification/2010/04/19/ntlm-keys-and-sundry-stuff/
for more details on NTLM keys

Signed-off-by: SARAT KUMAR BEHERA <beherasaratkumar@gmail.com>