always build with ntlmssp
Previously we would not build with ntlmssp if we had libkrb5.
Change this so that we always build with our ntlmssp implementation
and allow the user to select authentication mechanism through either the
URL argument or via an API call.

The reason for this is so that we will be able to use either krb5 or ntlmssp
from the same binary and still have the option to use sign or seal.
(the krb5 ntlmssp mech plugin does not export the session key so that
ntlm implementation can not be used for sign/seal).

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>