sync: don't access freed memory for cancelled PDUs
If we fail wait_for_reply in the sync API it means that the the underlying
SMB2 PDU might still be in fligth and our sync callback will be invoked later.

Solve this by not freeing the cb_data structure on wait_for_reply failures
and instead just flag the status as CANCELLED.
Later when the callback is invoked, if status is CANCELLED it means we should
just free cb_data and return.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>