dcerpc: massive updates to dcerpc and lsa
Make it possible to both encode and decode request and response messages
thus making it possible to do the marshalling a server would need
and not just client only marshalling (i.e. encode requests/decode replies)

Add a test program that tests encoding/decoding of the LSA function
and fix a lot of bugs in the dcerpc layer now that we try to do marshalling
both ways for all requests/replies.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>