NTLMSSP crash when we fail to parse an NTLMSSP challenge message
Check that we actually got an NTLMSSP packet back from the server
and not it upgrading to NTLMSSP wrapped in spnego.

Additionally add bounds checks to where we read fields from the NTLM buffer
when parsing the challenge packet from the server so that we do not read
beyond the end of the buffer.

This fixes segfaults when we could not parse a response correctly.

Reported-by: Marcus Roma <marcus.roma@advenica.com>
Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>