dcerpc: more refactoring of dcerpc making it less of akludge
goal is to have coder functions that can either marshall or unmarshall
depending on ->direction
and to handle all three types of pointers and alignment
based on conformance data sizes as well as maximum field alignment
of fields inside aggregate types (structs are aligned according to alignment
of the largest alignment of any immediately embedded fields)

Conformance data such as pointer sizes, arm-selector for unions and
max/offset/current sizes for conforming and varying arrays
all affect alignment so we need to refactor properly to make it possible to
have an easy to build ILD compiler for these protocols.
(negotiated transfer syntaxes such as NDR vs NDR64 also affect alignment.
 so, ...   but hey, this will be the third dcerpc/ndr implementation I have
written so better do it right this time so I never need to do it again.)

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>