srvsvc: add missing struct SHARE_INFO_STRUCT
The IDL we had for NetShareEnum was a bit busted but kludged to
work.
Add structure and coder for the missing SHARE_INFO_STRUCT
from MS-SRVC
Coder has a kludge to fix alignment because of the pointer in the embedded
union means it should be aligned at 64bits when NDR64 is used
and not at 32 bits (which the coder for the Level implicitely guarantees
anyway).

The full fix for this will come later when we always run the coder twice
for every aggregate.
An initial run to just decode the conformance data, and also collect
alignment information
and a second run where we apply the alignment and then process all the
actual data,

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>