dcerpc srvsvc: minor tweak
Not sure what way to handle when you have things like

int xxx
switch_is(xxx) foo_union

and automatically pass the outer xxx to the array arm in
foo_union.
For now have to specify both in the structures before encoding
it, but it would be nice to have it done automatically
and for decoding it would be nice to automatically
verify that both instances of xxx read from the iov
are actually the same value.
This is so much easier to handle when you have an idl compiler
that can automatically inject the code you need. But I
will have to do with what I have for now.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>