tests: add tests to simulate socket failures and check with valgrind for leaks
The test program prog_ls does leak in this situation so we need to fix that
and check if it is the test program or libsmb2 that leaks.
Then we can run tis type of tests on all binaries we build.

Valgrind logs on failure are in tests/valgrind.out

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>