open/opendir: Start converting to _async_pdu() api
Add new API for _open_async and _opendir_async
where we return a pdu pointer to the application.
The application can now free the pdu if it no longer
wants to wait for the reply from the server.
When the pdu is freed it is also removed from the
out an wait lists and the associated callback will
no longer be invoked even if the server does respond at a
later time.

Update the synchronous open and opendir calls to use this new api
which simplifies managing and releasing memopry during
failure paths.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>