Handle cases where we receive responses to cancelled PDUs
From 3818a74ae we now have the capability to cancel PDUs in flight.
Update the read_from_socket() to handle this case properly.
In this case we may receive a response from the server to a PDU we
have already cancelled and are no longer waiting for.
Instead of treating this as a protocol error, just threat the rest
of the "no longer wanted" reply as if was basically padding
but no associated pdu and no callbacks to invoke.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>