Stop prog_ls blowing the stack on startup on Linux
tests/prog_ls.c: Apparently Linux's dlsym uses calloc in its implementation, presenting a challenge for anyone hoping to shim calloc using dlsym.  https://stackoverflow.com/questions/7910666/problems-with-ld-preload-and-calloc-interposition-for-certain-executables doesn't just tell us that, though, it tells us that we can safely return NULL to dlsym to break the recursion.  I haven't verified that we still take the expected paths with the alloc_fail cleverness but this gets prog_ls up and working.