Test the D163336 overdrawn credit issue
tests/metastat-0202-censored.c: Amit Brahmbhatt's reproducer with just two changes from the original:
* the SMB2 dialect hard-coded to 2.0.2, for ease of testing particularly against the likes of Windows Server 2008 R2, where I know of no documented interface for changing that on the Windows end
* the useful local example credentials elided for teh security

tests/test_0400_overdrawn_0202.sh: A driver script for the most recent evolution of the invocation of that.  H/T https://stackoverflow.com/a/356154/18096 for the way to collect the exit statuses, which looks like Bash to me.  Having joined battle with prog_ls, I use that to find a filename on the server to stat against, making the assumption that there is one.

When I ran this standalone just now against Windows Server 2008 R2, it passed 10 out of 10 attempts with the suggested fix and failed 8 out of 10 attempts without the fix.

tests/Makefile.am: I need to add that to the list of programs needing compilation.  I like my Makefile lists to have one thing per line, in alphabetical order, to ease navigation of the history.